Authors blog page (#74)
* [docs][ui]: remove back ticks from code tag in mdx

* [author-blogs]: init author blogs page

* [author-blogs]: finished blogs and render authors

* [author-blogs]: refactor author card component and add at blogs end

* [author-blogs]: using the author card component

* [author-blogs]: responsive blogs page

* [author-blogs]: add nextbase blogs text

* [author-blogs]: responsive improvements

* Apply suggestions from code review

* Apply suggestions from code review

* [author-blogs]: feature

-- Make author optional on blogs create/update page. -- Handle on public page if author exist then only show author card -- show create blog button with out creating author profiles.

---------

+++ src/app/(dynamic-pages)/(authenticated-pages)/app_admin/blog/actions.ts
@@ -118,32 +118,34 @@ export const updateBlogPost = async (
     .eq('id', postId)
     .select('*')
     .single();
-
   if (error) {
     throw error;
   }
-
-  const { data: oldAuthors, error: oldAuthorsError } = await supabaseAdminClient
-    .from('internal_blog_author_posts')
-    .select('*')
-    .eq('post_id', postId);
-  if (oldAuthorsError) {
-    throw oldAuthorsError;
-  }
-
-  for (const oldAuthor of oldAuthors) {
-    const { error: deleteError } = await supabaseAdminClient
-      .from('internal_blog_author_posts')
-      .delete()
-      .eq('author_id', oldAuthor.author_id)
-      .eq('post_id', postId);
-    if (deleteError) {
-      throw deleteError;
+  if (authorId) {
+    const { data: oldAuthors, error: oldAuthorsError } =
+      await supabaseAdminClient
+        .from('internal_blog_author_posts')
+        .select('*')
+        .eq('post_id', postId);
+    if (oldAuthorsError) {
+      throw oldAuthorsError;
     }
+
+    for (const oldAuthor of oldAuthors) {
+      const { error: deleteError } = await supabaseAdminClient
+        .from('internal_blog_author_posts')
+        .delete()
+        .eq('author_id', oldAuthor.author_id)
+        .eq('post_id', postId);
+      if (deleteError) {
+        throw deleteError;
+      }
+    }
+
+    // assign new author to the post
+    await assignBlogPostToAuthor(authorId, postId);
   }
 
-  // assign new author to the post
-  await assignBlogPostToAuthor(authorId, postId);
   await updateBlogTagRelationships(data.id, tagIds);
 
   revalidatePath('/');
